+++ translator/test_infra/sol/if_fn.sol
@@ -8,7 +8,7 @@ library IfFn {
     // # test_1 (true, false, true) 1
     // # test_1 (true, true, false) 2
     // # test_1 (false, false, false) 0
-    function test_1(bool a, bool b, bool c) public pure returns (uint){
+    function test_1(bool a, bool b, bool c) public pure returns (uint128){
         if (a) {
             if(b){
                 if(c){
@@ -25,7 +25,7 @@ library IfFn {
 
     // # test_2 (true, 1, 2) 1
     // # test_2 (false, 1, 2) 3
-    function test_2(bool a, uint b, uint c) public pure returns (uint){
+    function test_2(bool a, uint128 b, uint128 c) public pure returns (uint128){
         if (a) {
             return b;
         }
@@ -35,7 +35,7 @@ library IfFn {
 
     // # test_3 (1, 2) 2
     // # test_3 (256, 0) 256
-    function test_3(uint a, uint b) public pure returns (uint){
+    function test_3(uint128 a, uint128 b) public pure returns (uint128){
         return  (a > b ? a : b);
     }
 }
