+++ translator/test_infra/sol/uniswap/l_SafeMath.sol
@@ -7,15 +7,15 @@ pragma solidity >=0.5.16;
 // a library for performing overflow-safe math, courtesy of DappHub (https://github.com/dapphub/ds-math)
 
 library SafeMath {
-    function add(uint x, uint y) internal pure returns (uint z) {
+    function add(uint128 x, uint128 y) internal pure returns (uint128 z) {
         require((z = x + y) >= x, 'ds-math-add-overflow');
     }
 
-    function sub(uint x, uint y) internal pure returns (uint z) {
+    function sub(uint128 x, uint128 y) internal pure returns (uint128 z) {
         require((z = x - y) <= x, 'ds-math-sub-underflow');
     }
 
-    function mul(uint x, uint y) internal pure returns (uint z) {
+    function mul(uint128 x, uint128 y) internal pure returns (uint128 z) {
         require(y == 0 || (z = x * y) / y == x, 'ds-math-mul-overflow');
     }
 }
